LPS-206371 Merger the class FeatureFlagBatchEngineUnitProcessor with the class BatchEngineUnitProcessorImpl
